+++ src/main/ruby/truffleruby/core/weakmap.rb
@@ -9,29 +9,11 @@
 # GNU Lesser General Public License version 2.1.
 
 module ObjectSpace
-  # WeakMap uses identity comparison semantics. The implementation assumes that the Java representation of objects
-  # do compare (equals() and hashCode()) using object identity. This is the case for instances of RubyDynamicObject.
-  #
-  # However, results are unspecified if used with instances of TruffleObject that override equals() and
-  # hashCode().
-  #
-  # Note that, following MRI, we disallow immediate (primitive) values and frozen objects as keys or value of
-  # WeakMap. We could easily transcend this limitation as we do not modify objects like MRI does (it sets a finalizer).
-  # However, keeping the limitation enables the above assumption of identity comparison.
+  # WeakMap uses Ruby identity semantics to compare and hash keys.
   class WeakMap
     include Enumerable
 
-    private def check_key_or_value(kv, type)
-      if Primitive.immediate_value?(kv)
-        raise ArgumentError, "WeakMap #{type} can't be an instance of #{kv.class}"
-      elsif Truffle::KernelOperations.value_frozen?(kv)
-        raise FrozenError, "WeakMap #{type} can't be a frozen object"
-      end
-    end
-
     def []=(key, value)
-      check_key_or_value(key, 'key')
-      check_key_or_value(value, 'value')
       Primitive.weakmap_aset(self, key, value)
     end