+++ CHANGELOG.md
@@ -69,6 +69,7 @@ Compatibility:
 * Implemented `Kernel#raise` `cause` parameter.
 * Improved compatibility of `Signal.trap` and `Kernel#trap` (#2287, @chrisseaton).
 * Implemented `GC.stat(:total_allocated_objects)` as `0` (#2292, @chrisseaton).
+* `ObjectSpace::WeakMap` now supports immediate and frozen values as both keys and values (#2267).
 
 Performance: